requirements.txt
Django==6.0
djangorestframework==3.16.1
drf_yasg==1.21.11
environ==1.0
jsonschema==4.25.1
Requests==2.32.5
sendgrid==6.12.5

#与前端，跨域相关配置
django-cors-headers==4.4.0